:bug: fix: Function didnt run because of an awaitModal error and discord limit, so it was split into two to fix it
